+++ components/examples/ButtonExamples.tsx
@@ -1,6 +1,6 @@
 import React, { useState } from 'react';
 import { View, Text } from 'react-native';
-import CustomButton from '@/components/common/Button/CustomButton';
+import { CustomButton } from '@/components';
 
 export default function ButtonExamples() {
   return (
